refactor : racing 수정
